Move more code to stats context
