+++ rest/server_context.go
@@ -13,14 +13,11 @@ import (
 	"bytes"
 	"encoding/json"
 	"errors"
-	"expvar"
 	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/url"
-	"os"
 	"regexp"
-	"runtime"
 	"strconv"
 	"strings"
 	"sync"
@@ -30,7 +27,6 @@ import (
 	"github.com/couchbase/go-couchbase"
 	"github.com/couchbase/sync_gateway/base"
 	"github.com/couchbase/sync_gateway/db"
-	"github.com/elastic/gosigar"
 	pkgerrors "github.com/pkg/errors"
 )
 
@@ -989,7 +985,7 @@ func (sc *ServerContext) logStats() error {
 
 	AddGoRuntimeStats()
 
-	if err := sc.addGoSigarStats(); err != nil {
+	if err := sc.statsContext.addGoSigarStats(); err != nil {
 		base.Warnf(base.KeyAll, "Error getting sigar based system resource stats: %v", err)
 	}
 
@@ -1017,165 +1013,6 @@ func (sc *ServerContext) logStats() error {
 
 }
 
-func (sc *ServerContext) addGoSigarStats() error {
-
-	if err := sc.addProcessCpuPercentage(); err != nil {
-		return err
-	}
-
-	if err := sc.addProcessMemoryPercentage(); err != nil {
-		return err
-	}
-
-	return nil
-
-}
-
-func (sc *ServerContext) addProcessCpuPercentage() error {
-
-	statsResourceUtilization := base.StatsResourceUtilization()
-
-	// Calculate the cpu percentage for the process
-	cpuPercentUtilization, err := sc.calculateProcessCpuPercentage()
-	if err != nil {
-		return err
-	}
-
-	// Record stat
-	statsResourceUtilization.Set(base.StatKeyProcessCpuPercentUtilization, base.ExpvarFloatVal(cpuPercentUtilization))
-
-	return nil
-
-}
-
-func (sc *ServerContext) addProcessMemoryPercentage() error {
-
-	statsResourceUtilization := base.StatsResourceUtilization()
-
-	// Calculate the memory utilization percentage for the process
-	memoryPercentUtilization, err := sc.calculateProcessMemoryPercentage()
-	if err != nil {
-		return err
-	}
-
-	// Record stat
-	statsResourceUtilization.Set(base.StatKeyProcessMemoryPercentUtilization, base.ExpvarFloatVal(memoryPercentUtilization))
-
-	return nil
-
-}
-
-// Calculate the percentage of CPU used by this process over the sampling time specified in statsLogFrequencySecs
-//
-// Based on the accepted answer by "caf" in: https://stackoverflow.com/questions/1420426/how-to-calculate-the-cpu-usage-of-a-process-by-pid-in-linux-from-c
-//
-// This has a minor variation, and rather than directly:
-//
-// - Collect stats sample
-// - Sleep for sample time (1s in the SO post, but that's fairly arbitrary)
-// - Collect stats sample again
-// - Calculate process cpu percentage over sample period
-//
-// It uses the same time.Ticker as for the other stats collection, and stores a previous stats sample and compares
-// the current against the previous to calcuate the cpu percentage.  If it's the first time it's invoked, there
-// won't be a previous value and so it will record 0.0 as the cpu percentage in that case.
-func (sc *ServerContext) calculateProcessCpuPercentage() (cpuPercentUtilization float64, err error) {
-
-	// Get current value
-	currentSnapshot, err := newCpuStatsSnapshot()
-	if err != nil {
-		return 0, err
-	}
-
-	// Is there a previous value?  If not, store current value as previous value and don't log a stat
-	if sc.statsContext.cpuStatsSnapshot == nil {
-		sc.statsContext.cpuStatsSnapshot = currentSnapshot
-		return 0, nil
-	}
-
-	// Otherwise calculate the cpu percentage based on current vs previous
-	prevSnapshot := sc.statsContext.cpuStatsSnapshot
-
-	// The delta in user time for the process
-	deltaUserTimeJiffies := float64(currentSnapshot.procUserTimeJiffies - prevSnapshot.procUserTimeJiffies)
-
-	// The delta in system time for the process
-	deltaSystemTimeJiffies := float64(currentSnapshot.procSystemTimeJiffies - prevSnapshot.procSystemTimeJiffies)
-
-	// The combined delta of user + system time for the process
-	deltaSgProcessTimeJiffies := deltaUserTimeJiffies + deltaSystemTimeJiffies
-
-	// The delta in total time for the machine
-	deltaTotalTimeJiffies := float64(currentSnapshot.totalTimeJiffies - prevSnapshot.totalTimeJiffies)
-
-	// Calculate the CPU usage percentage for the SG process
-	cpuPercentUtilization = 100 * deltaSgProcessTimeJiffies / deltaTotalTimeJiffies
-
-	// Store the current values as the previous values for the next time this function is called
-	sc.statsContext.cpuStatsSnapshot = currentSnapshot
-
-	return cpuPercentUtilization, nil
-
-}
-
-func (sc *ServerContext) calculateProcessMemoryPercentage() (memoryPercentUtilization float64, err error) {
-
-	pid := os.Getpid()
-
-	procMem := gosigar.ProcMem{}
-	if err := procMem.Get(pid); err != nil {
-		return 0, nil
-	}
-
-	totalMem := gosigar.Mem{}
-	if err := totalMem.Get(); err != nil {
-		return 0, nil
-	}
-
-	memoryPercentUtilization = float64(procMem.Resident) / float64(totalMem.Total)
-
-	return memoryPercentUtilization, nil
-
-}
-
-func AddGoRuntimeStats() {
-
-	statsResourceUtilization := base.StatsResourceUtilization()
-
-	// Num goroutines
-	statsResourceUtilization.Set(base.StatKeyNumGoroutines, base.ExpvarIntVal(runtime.NumGoroutine()))
-
-	recordGoroutineHighwaterMark(statsResourceUtilization, uint64(runtime.NumGoroutine()))
-
-	// Read memstats (relatively expensive)
-	memstats := runtime.MemStats{}
-	runtime.ReadMemStats(&memstats)
-
-	// Sys
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsSys, base.ExpvarUInt64Val(memstats.Sys))
-
-	// HeapAlloc
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsHeapAlloc, base.ExpvarUInt64Val(memstats.HeapAlloc))
-
-	// HeapIdle
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsHeapIdle, base.ExpvarUInt64Val(memstats.HeapIdle))
-
-	// HeapInuse
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsHeapInUse, base.ExpvarUInt64Val(memstats.HeapInuse))
-
-	// HeapReleased
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsHeapReleased, base.ExpvarUInt64Val(memstats.HeapReleased))
-
-	// StackInuse
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsStackInUse, base.ExpvarUInt64Val(memstats.StackInuse))
-
-	// StackSys
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsStackSys, base.ExpvarUInt64Val(memstats.StackSys))
-
-	// PauseTotalNs
-	statsResourceUtilization.Set(base.StatKeyGoMemstatsPauseTotalNs, base.ExpvarUInt64Val(memstats.PauseTotalNs))
-
-}
 
 // Updates stats that are more efficient to calculate at stats collection time
 func (sc *ServerContext) updateCalculatedStats() {
@@ -1187,26 +1024,6 @@ func (sc *ServerContext) updateCalculatedStats() {
 
 }
 
-// Record Goroutines high watermark into expvars
-func recordGoroutineHighwaterMark(stats *expvar.Map, numGoroutines uint64) (maxGoroutinesSeen uint64) {
-
-	maxGoroutinesSeen = atomic.LoadUint64(&base.MaxGoroutinesSeen)
-
-	if numGoroutines > maxGoroutinesSeen {
-
-		// Clobber existing values rather than attempt a CAS loop. This stat can be considered a "best effort".
-		atomic.StoreUint64(&base.MaxGoroutinesSeen, numGoroutines)
-
-		if stats != nil {
-			stats.Set(base.StatKeyGoroutinesHighWatermark, base.ExpvarUInt64Val(maxGoroutinesSeen))
-		}
-
-		return numGoroutines
-	}
-
-	return maxGoroutinesSeen
-
-}
 
 // For test use
 func (sc *ServerContext) Database(name string) *db.DatabaseContext {
